Update configure-tempo-data-source.md
I was using Loki as Logs datasource and adding 'uid' for Loki datasources didn't work and even the datasource wasn't showing up.  

Upon experimenting, I gave the 'name' of Loki datasource as datasourceUid in tracesToLogsV2 section and it worked. However, it took a lot of time for me to figure it out as the doc said it is 'uid' value of datasource instead of 'name'